Gem Stone Problem
John has discovered various rocks. Each rock is composed of various
elements, and each element is represented by a lowercase latin letter
from 'a' to 'z'. An element can be present multiple times in a rock. An
element is called a 'gem-element' if it occurs at least once in each of
the rocks.

Given the list of rocks with their compositions, display the number of
gem-elements that exist in those rocks.